refactor, port tests to mocha
