+++ .nycrc
@@ -1,0 +1,18 @@
+{
+  "extension": [
+    ".ts"
+  ],
+  "include" :[
+    "src/**/*.ts"
+  ],
+  "exclude": [
+    "dist",
+    "test",
+    "**/*.d.ts"
+  ],
+  "reporter": [
+    "html",
+    "lcov",
+    "text-summary"
+  ]
+}
