+++ package.json
@@ -6,7 +6,9 @@
   "scripts": {
     "build": "tsc -p tsconfig.json",
     "watch": "tsc -p tsconfig.json -w",
-    "test": "ts-node node_modules/tape/bin/tape tests/**/*.ts"
+    "test": "mocha -r ts-node/register/transpile-only --full-trace 'test/**/*.ts'",
+    "test:watch": "mocha -r ts-node/register/transpile-only --watch --watch-files src/*.ts,test/**/*.ts --full-trace 'test/**/*.ts'",
+    "coverage": "nyc npm test"
   },
   "files": [
     "dist",
@@ -28,8 +30,10 @@
   "homepage": "https://github.com/thomn/classnames#readme",
   "devDependencies": {
     "@types/node": "^14.14.10",
-    "tape": "^5.0.1",
     "ts-node": "^9.1.0",
     "typescript": "^4.1.2"
+    "@types/mocha": "^9.0.0",
+    "mocha": "^9.1.3",
+    "nyc": "^15.1.0",
   }
 }
